# docs/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TenantModuleUsageItemResponse": {
        "properties": {
          "at_limit": {
            "default": false,
            "title": "At Limit",
            "type": "boolean"
          },
          "limit_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit Source"
          },
          "max_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Units"
          },
          "module_key": {
            "title": "Module Key",
            "type": "string"
          },
          "module_name": {
            "title": "Module Name",
            "type": "string"
          },
          "remaining_units": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining Units"
          },
          "unlimited": {
            "default": false,
            "title": "Unlimited",
            "type": "boolean"
          },
          "used_units": {
            "title": "Used Units",
            "type": "integer"
          }
        },
        "required": [
          "module_name",
          "module_key",
          "used_units"
        ],
        "title": "TenantModuleUsageItemResponse",
        "type": "object"
      },
...
      "TenantModuleUsageSummaryResponse": {
        "properties": {
          "billing_in_grace": {
            "default": false,
            "title": "Billing In Grace",
            "type": "boolean"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/TenantModuleUsageItemResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "integer"
          },
          "tenant_plan_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Plan Code"
          },
          "tenant_slug": {
            "title": "Tenant Slug",
            "type": "string"
          },
          "tenant_status": {
            "title": "Tenant Status",
            "type": "string"
          },
          "total_modules": {
            "title": "Total Modules",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "message",
          "tenant_id",
          "tenant_slug",
          "tenant_status",
          "total_modules",
          "data"
        ],
        "title": "TenantModuleUsageSummaryResponse",
        "type": "object"
      },
...
      "TenantUserCreateRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "password": {
            "title": "Password",
            "type": "string"
          },
          "role": {
            "default": "operator",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "email",
          "password"
        ],
        "title": "TenantUserCreateRequest",
        "type": "object"
      },
...
      "TenantUserDetailResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TenantUsersItemResponse"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requested_by": {
            "$ref": "#/components/schemas/TenantUserContextResponse"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "requested_by",
          "data"
        ],
        "title": "TenantUserDetailResponse",
        "type": "object"
      },
      "TenantUserMutationResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TenantUsersItemResponse"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requested_by": {
            "$ref": "#/components/schemas/TenantUserContextResponse"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message",
          "requested_by",
          "data"
        ],
        "title": "TenantUserMutationResponse",
        "type": "object"
      },
      "TenantUserStatusUpdateRequest": {
        "properties": {
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          }
        },
        "required": [
          "is_active"
        ],
        "title": "TenantUserStatusUpdateRequest",
        "type": "object"
      },
      "TenantUserUpdateRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "full_name",
          "email",
          "role"
        ],
        "title": "TenantUserUpdateRequest",
        "type": "object"
      },
...
      "TenantUsersResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TenantUsersItemResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requested_by": {
            "$ref": "#/components/schemas/TenantUserContextResponse"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "success",
          "message",
          "requested_by",
          "total",
          "data"
        ],
        "title": "TenantUsersResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Platform Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/platform/tenants/{tenant_id}/module-usage": {
      "get": {
        "operationId": "get_tenant_module_usage_platform_tenants__tenant_id__module_usage_get",
        "parameters": [
          {
            "in": "path",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "title": "Tenant Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantModuleUsageSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tenant Module Usage",
        "tags": [
          "platform-tenants"
        ]
      }
    },
...
    "/tenant/users": {
      "get": {
        "operationId": "tenant_users_tenant_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUsersResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Tenant Users",
        "tags": [
          "Tenant Protected"
        ]
      },
      "post": {
        "operationId": "tenant_create_user_tenant_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUserCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUserMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenant Create User",
        "tags": [
          "Tenant Protected"
        ]
      }
    },
...
    "/tenant/users/{user_id}": {
      "get": {
        "operationId": "tenant_user_detail_tenant_users__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUserDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenant User Detail",
        "tags": [
          "Tenant Protected"
        ]
      },
      "put": {
        "operationId": "tenant_update_user_tenant_users__user_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUserMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenant Update User",
        "tags": [
          "Tenant Protected"
        ]
      }
    },
    "/tenant/users/{user_id}/status": {
      "patch": {
        "operationId": "tenant_update_user_status_tenant_users__user_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantUserStatusUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenantUserMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tenant Update User Status",
        "tags": [
          "Tenant Protected"
        ]
      }
    },
...
  }
}